services/speech-to-text/pyproject.toml
[project]
name = "speech-to-text"
version = "0.1.0"
description = "FastAPI microservice for speech-to-text transcription"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "nemo_toolkit[asr]>=2.0.0",
    "python-multipart>=0.0.9",
]
